requirement.txt
openai>=1.0.0
googlemaps
python-dotenv
